Add logging class (#105)
* ignore additional clasp json files

* extract logging to distinct class
