+++ lib/main.ts
@@ -4,11 +4,12 @@
 
 import FileService from './FileService';
 import GDriveService from './GDriveService';
-import { Util } from './Util';
+import Util from './Util';
 import Properties from './Properties';
 import Timer from './Timer';
 import TriggerService from './TriggerService';
 import ErrorMessages from './ErrorMessages';
+import Logging from './util/Logging';
 
 /**
  * Copy folders and files from source to destination.
@@ -104,7 +105,7 @@ export function copy(): void {
       try {
         fileList = gDriveService.getFiles(query, properties.pageToken);
       } catch (e) {
-        Util.log({ ss, status: Util.composeErrorMsg(e) });
+        Logging.log({ ss, status: Util.composeErrorMsg(e) });
       }
       if (!fileList) {
         console.log('fileList is undefined. currFolder:', currFolder);
